pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Jesse Zhao <zhao.j.jesse@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
openai = "^1.51.0"
pytest = "^8.3.3"
streamlit = "^1.39.0"
ipdb = "^0.13.13"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
